requirements.txt
openai
uagents
uagents-core
python-dotenv
pymongo
fastapi
uvicorn
langgraph

